% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{dextergui}
\alias{dextergui}
\title{Start dextergui}
\usage{
dextergui(dbpath = NULL, wd = getwd(), roots = NULL)
}
\arguments{
\item{dbpath}{path to a dexter project database
or NULL, in which case you can select a project after starting dextergui}

\item{wd}{where dextergui first looks for and saves project files, defaults to current working directory.}

\item{roots}{volumes or paths on your hard drive available for opening and saving project files. Set NULL
to use all accessible volumes}
}
\description{
Opens a shiny application providing a graphical user interface to dexter.
}
\details{
The best results are achieved when the gui is opened in a browser (Chrome, Brave, FireFox). Somewhat
less aesthetically pleasing results are achieved in Internet Explorer. The Edge browser is not supported at this time.

The RStudio browser does not currently support downloads of plots and tables. Starting the gui
in your default browser automatically can be achieved in several ways. One way, shown below, 
is to set the \code{shiny.launch.browser} option to \code{TRUE}.
}
\examples{
\dontrun{

options(shiny.launch.browser = TRUE)


dextergui()

}



}
