% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{dextergui}
\alias{dextergui}
\title{Start dextergui}
\usage{
dextergui(dbpath = NULL, wd = getwd())
}
\arguments{
\item{dbpath}{path to a dexter project database
or NULL, in which case you can select a project after starting dextergui}

\item{wd}{where dextergui looks for and saves project files, defaults to current working directory. 
Within the gui you can move to subdirectories of \code{wd} but not further up the tree than \code{wd}.}
}
\description{
Opens a shiny application providing a graphical user interface to dexter.
}
\details{
The best results are achieved when the gui is opened in a browser (Chrome, Brave, FireFox). Somewhat
less aesthetically pleasing results are achieved in Internet Explorer. The Edge browser is not supported.

The native RStudio browser does not currently support downloads of plots and tables. Starting the gui
in the browser automatically can be achieved in multiple ways. One way is to set the shiny.launch.browser option to true.
}
\examples{
\dontrun{

options(shiny.launch.browser = TRUE)


dextergui()

}



}
