% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve.R
\name{dextergui}
\alias{dextergui}
\title{Start dextergui}
\usage{
dextergui(dbpath = NULL, wd = getwd())
}
\arguments{
\item{dbpath}{path to a dexter project database
or NULL, in which case you can select a project after starting dextergui}

\item{wd}{where dextergui looks for and saves project files, defaults to current working directory. 
Within the gui you can move to subdirectories of \code{wd} but not further up the tree than \code{wd}.}
}
\description{
Opens a shiny application providing a graphical user interface to dexter.
}
